Simplify effective schema bootstrap startup task
